requirements.txt
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.1
django-filter==2.4.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.1.0
psycopg2==2.8.6
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
